Refactor to make the code follow the MVI principles
MVI:

Model-View-Intent (MVI) is an architectural pattern that is based on
unidirectional data flow and immutable state. It's often used in
reactive programming. Here's a short summary of its principles:

Model: The Model represents the state of the application. It's an
       immutable data structure that contains all the necessary
       information to render the View. Any change in the application
       results in a new Model.

View: The View is responsible for rendering the Model and emitting
      Intents. It's a passive interface that displays the Model and
      routes user interactions to the Intent.

Intent: The Intent represents an intention or desire to perform an
        action, usually triggered by user interactions. It's a
        simple object that describes what the user wants to do.

The data flow in MVI is unidirectional: the View emits Intents,
the ViewModel processes these Intents and updates the Model,
and the View renders the new Model. This cycle repeats for every
user interaction, resulting in a predictable and easy-to-follow
data flow.

How MVI is implemented in this Application:

The HomeScreen is the main screen of the application. It is a
composable function that takes a MainViewModel as a parameter.
The content of the body is determined by the current tab,
which is a state variable.

The HomeScreen function defines several state variables that are
collected from the ViewModel, including apiError, loading,
nsoPackages, nsoDevices, nsoAlarms, and refresh. These state
variables are observed and any changes to them will cause a
recompose to occur of the appropriate parts of the UI.

For the PackagesScreen, DevicesScreen, and AlarmsScreen, before
they are displayed, an intent is sent to the ViewModel to show
the packages, devices, or alarms, respectively. This will trigger
the ViewModel to enable fetching data and/or update the state,
changes which then is observed by the HomeScreen.